Update structure to match refactor branch
