+++ tests/test_github_query.py
@@ -2,7 +2,7 @@ from typing import Any, Literal
 import json
 import pytest
 
-import github_query
+from src import conversion_logic, queries
 
 @pytest.fixture
 def pr_api_output() -> list[dict[str, Any]]:
@@ -115,13 +115,13 @@ def csv_string_empty() -> Literal['']:
 # Get PR Label test-cases
 
 def test_get_labels(pr_api_output):
-    labels = github_query.get_labels(pr_data=pr_api_output)
+    labels = conversion_logic.filter_unique_labels(pr_data=pr_api_output)
 
     assert isinstance(labels, list)
     assert set(labels) == {"bugfix", "enhancement"}
 
 def test_get_labels_missing_input(pr_api_output_missing_label):
-    labels = github_query.get_labels(pr_data=pr_api_output_missing_label)
+    labels = conversion_logic.filter_unique_labels(pr_data=pr_api_output_missing_label)
 
     assert labels == []
 
@@ -129,22 +129,22 @@ def test_get_labels_missing_input(pr_api_output_missing_label):
 # Convert repo label list
 
 def test_csv_string_to_list_spaces(csv_string_spaces):
-    string_list = github_query.csv_string_to_list(csv_string_spaces)
+    string_list = conversion_logic.csv_string_to_list(csv_string_spaces)
 
     assert string_list == ["bugfix", "enhancement", "feature"]
 
 def test_csv_string_to_list_no_spaces(csv_string_no_spaces):
-    string_list = github_query.csv_string_to_list(csv_string_no_spaces)
+    string_list = conversion_logic.csv_string_to_list(csv_string_no_spaces)
 
     assert string_list == ["bugfix", "enhancement", "feature"]
 
 def test_csv_string_to_list_no_comma(csv_string_no_comma):
-    string_list = github_query.csv_string_to_list(csv_string_no_comma)
+    string_list = conversion_logic.csv_string_to_list(csv_string_no_comma)
 
     assert string_list == ["bugfix"]
 
 def test_csv_string_to_list_empty(csv_string_empty):
-    string_list = github_query.csv_string_to_list(csv_string_empty)
+    string_list = conversion_logic.csv_string_to_list(csv_string_empty)
 
     assert string_list == []
 
@@ -152,27 +152,27 @@ def test_csv_string_to_list_empty(csv_string_empty):
 # Version Increment test-cases
 
 def test_get_version_increment_patch(minor_bump, patch_bump, pr_labels_bug):
-    increment = github_query.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_bug)
+    increment = conversion_logic.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_bug)
 
     assert increment == "patch"
 
 def test_get_version_increment_minor(minor_bump, patch_bump, pr_labels_enhancement):
-    increment = github_query.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_enhancement)
+    increment = conversion_logic.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_enhancement)
 
     assert increment == "minor"
 
 def test_get_version_increment_wrong_labels(minor_bump, patch_bump, pr_labels_wrong_labels):
-    increment = github_query.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_wrong_labels)
+    increment = conversion_logic.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_wrong_labels)
 
     assert increment == ""
 
 def test_get_version_increment_none(minor_bump, patch_bump, pr_labels_none):
-    increment = github_query.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_none)
+    increment = conversion_logic.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_none)
 
     assert increment == ""
 
 def test_get_version_increment_empty_list(minor_bump, patch_bump, pr_labels_empty_list):
-    increment = github_query.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_empty_list)
+    increment = conversion_logic.get_version_increment(patch_bump_list=patch_bump, minor_bump_list=minor_bump, pr_label_list=pr_labels_empty_list)
 
     assert increment == ""
 
@@ -181,7 +181,7 @@ def test_get_version_increment_empty_list(minor_bump, patch_bump, pr_labels_empt
 
 def test_changer_pert_label(merged_pr_samples: dict[str, str]) -> None:
     changelog_labels: list[str] = ["type: bug", "type: enhancement", "type: maintenance"]
-    filtered_results: list[Changelog] = github_query.filter_changes_per_label(pr_data=merged_pr_samples, changelog_label_list=changelog_labels)
+    filtered_results: list[Changelog] = conversion_logic.filter_changes_per_label(pr_data=merged_pr_samples, changelog_label_list=changelog_labels)
 
     for result in filtered_results:
         for label in result.labels:
